pyproject.toml
[project]
name = "acp-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "acp-sdk>=1.0.0",
    "crewai>=0.140.0",
    "crewai-tools>=0.49.0",
    "langgraph>=0.5.1",
    "openai>=1.93.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
